+++ lib/f_core/device/sensor/c_temperature_sensor.cpp
@@ -15,7 +15,7 @@ sensor_value CTemperatureSensor::GetSensorValue(sensor_channel chan) const {
         return temperature;
     }
 
-    LOG_ERR("Invalid sensor channel (%d) called for temperature sensor", chan);
+    LOG_ERR("Invalid sensor channel (%d) called for temperature sensor", static_cast<int>(chan));
     k_oops();
     return {INT32_MIN, INT32_MIN};
 }
